change types for swangle and brake pressure

+++ DCM/Inc/lut_3d.h
@@ -21,9 +21,9 @@ kappaAndFx getKappaFxGlobalMaxAtDownforce(float downforce_N, uint8_t throttlePos
 
 float getMaxKappaCurrentState(motorLocation_t motor, bool assumeNoTurn);
 
-float getBrakeKappa(motorLocation_t motor, uint8_t brakePressurePsi_u8, float deadband);
+float getBrakeKappa(motorLocation_t motor, uint16_t brakePressurePsi_u8, float deadband);
 
-int32_t getBrakeMaxTorque_mNm(motorLocation_t motor, uint8_t brakePressurePsi_u8);
+int32_t getBrakeMaxTorque_mNm(motorLocation_t motor, uint16_t brakePressurePsi_u8);
 
 float test();
